# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MIU Export Hub API",
    "description": "Made in Uganda Export Hub \u2014 B2B export marketplace backend.\n\n**Authentication:** send JWTs in the `Authorization` header, not query params.\n- Protected routes: `Authorization: Bearer <access_token>`\n- `/auth/refresh` and `/auth/logout`: `Authorization: Bearer <refresh_token>`",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/buyer/onboarding": {
      "get": {
        "tags": [
          "buyer"
        ],
        "summary": "Get Onboarding",
        "operationId": "get_onboarding_api_v1_buyer_onboarding_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerOnboardingDraftResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/buyer/onboarding/company": {
      "put": {
        "tags": [
          "buyer"
        ],
        "summary": "Onboarding Company",
        "operationId": "onboarding_company_api_v1_buyer_onboarding_company_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerCompanyStep"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/buyer/onboarding/contact": {
      "put": {
        "tags": [
          "buyer"
        ],
        "summary": "Onboarding Contact",
        "operationId": "onboarding_contact_api_v1_buyer_onboarding_contact_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerContactStep"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/buyer/onboarding/sourcing": {
      "put": {
        "tags": [
          "buyer"
        ],
        "summary": "Onboarding Sourcing",
        "operationId": "onboarding_sourcing_api_v1_buyer_onboarding_sourcing_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerSourcingStep"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/buyer/onboarding/documents": {
      "post": {
        "tags": [
          "buyer"
        ],
        "summary": "Onboarding Documents",
        "description": "Upload optional buyer verification documents (stub storage).",
        "operationId": "onboarding_documents_api_v1_buyer_onboarding_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_onboarding_documents_api_v1_buyer_onboarding_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/buyer/onboarding/submit": {
      "post": {
        "tags": [
          "buyer"
        ],
        "summary": "Onboarding Submit",
        "operationId": "onboarding_submit_api_v1_buyer_onboarding_submit_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/buyer/onboarding/status": {
      "get": {
        "tags": [
          "buyer"
        ],
        "summary": "Onboarding Status",
        "operationId": "onboarding_status_api_v1_buyer_onboarding_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerOnboardingStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_onboarding_documents_api_v1_buyer_onboarding_documents_post": {
        "properties": {
          "document_type": {
            "type": "string",
            "title": "Document Type"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "document_type",
          "file"
        ],
        "title": "Body_onboarding_documents_api_v1_buyer_onboarding_documents_post"
      },
...
      "BuyerCompanyStep": {
        "properties": {
          "company_name": {
            "type": "string",
            "minLength": 2,
            "title": "Company Name"
          },
          "country": {
            "type": "string",
            "minLength": 2,
            "title": "Country"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          }
        },
        "type": "object",
        "required": [
          "company_name",
          "country"
        ],
        "title": "BuyerCompanyStep"
      },
      "BuyerContactStep": {
        "properties": {
          "contact_name": {
            "type": "string",
            "minLength": 2,
            "title": "Contact Name"
          },
          "job_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Title"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "type": "object",
        "required": [
          "contact_name"
        ],
        "title": "BuyerContactStep"
      },
      "BuyerOnboardingDraftResponse": {
        "properties": {
          "step": {
            "type": "string",
            "title": "Step"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "onboarding_status": {
            "type": "string",
            "title": "Onboarding Status"
          }
        },
        "type": "object",
        "required": [
          "step",
          "payload",
          "onboarding_status"
        ],
        "title": "BuyerOnboardingDraftResponse"
      },
      "BuyerOnboardingStatusResponse": {
        "properties": {
          "onboarding_status": {
            "type": "string",
            "title": "Onboarding Status"
          },
          "verified_buyer": {
            "type": "boolean",
            "title": "Verified Buyer"
          },
          "admin_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Message"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          }
        },
        "type": "object",
        "required": [
          "onboarding_status",
          "verified_buyer"
        ],
        "title": "BuyerOnboardingStatusResponse"
      },
...
      "BuyerSourcingStep": {
        "properties": {
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories"
          },
          "target_markets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Markets"
          },
          "annual_import_volume": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annual Import Volume"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "BuyerSourcingStep"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "JWT access token. Example: Authorization: Bearer eyJhbGciOiJIUzI1NiIs...",
        "scheme": "bearer"
      },
      "BearerRefresh": {
        "type": "http",
        "description": "Refresh token for token rotation and logout.",
        "scheme": "bearer"
      }
    }
  }
}